renamed workingdirectory to shared, makes more sense
