+++ src/helper/shared.c
