+++ src/helper/shared.h
@@ -11,12 +11,6 @@ typedef struct _sharedmemory {
 	unsigned int wdSize;
 	short FLC;
 	char filename[45];
-
-/* Don't think we need these to be in shared memory
- *	unsigned int wdDepth;
- *	unsigned int wdSize;
- *	unsigned int wdOffset;
- */
 } sharedMemory;
 
 void dirAdd(char*);
